refactor: split out the bindings generation into multiple files
